requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.8.4
aioselenium==0.0.1
aiosignal==1.3.1
anthropic==0.3.11
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
black==23.3.0
bleach==5.0.1
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
clickhouse-connect==0.5.23
cohere==4.24
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
display==1.0.0
distlib==0.3.6
distro==1.8.0
duckdb==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
faiss-cpu==1.7.4
fastavro==1.8.2
fastjsonschema==2.16.2
feedparser==6.0.10
ffmpy==0.3.1
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
gast==0.4.0
geonamescache==1.6.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio_client==0.5.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.6.0
iniconfig==2.0.0
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.12.0
jedi==0.18.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
langchain==0.0.154
langdetect==1.0.9
libclang==16.0.6
linkify-it-py==2.0.2
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.4
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mwclient==0.10.1
mwparserfromhell==0.6.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
nodeenv==1.7.0
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.26.0
oauthlib==3.2.2
olefile==0.46
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.7
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandas-stubs==2.0.1.230501
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2023.6.26
platformdirs==2.6.2
pluggy==1.2.0
posthog==3.0.1
pretty-errors==1.2.25
prettytable==3.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
pydub==0.25.1
pyee==8.2.2
pygame==2.3.0
Pygments==2.14.0
pygpt4all==1.0.1
pygptj==1.0.10
pygwalker==0.1.6.1
pyllamacpp==1.0.6
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.0.9
pyppeteer==1.0.2
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.1
python-json-logger==2.0.4
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
replicate==0.11.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.0.1
rsa==4.9
schedule==1.2.0
scipy==1.10.0
seaborn==0.12.2
selenium==4.11.2
semantic-version==2.10.0
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.5.post1
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.26.1
streamlit==1.22.0
streamlit-ace==0.1.1
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torchvision==0.15.2
tornado==6.2
traitlets==5.8.1
trio==0.22.0
trio-websocket==0.9.2
types-pytz==2023.3.0.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uc-micro-py==1.0.2
unstructured==0.6.3
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
watchfiles==0.19.0
wcwidth==0.2.5
webcolors==1.12
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.3.7
widgetsnbextension==4.0.5
wrapt==1.14.1
wsproto==1.2.0
XlsxWriter==3.1.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
